==> Month 9/05 - Topic modeling with Gensim.ipynb <==
# Topic modeling with Gensim

Gensim's preprocessing tools can be great for extremely fast preprocessing, but this isn't what Gensim is really designed for.  Gensim is designed mostly around _unsupervised learning_ tasks in NLP.  Namely, topic modeling and text embeddings.  This notebook covers Gensim's topic modeling in _very_ brief detail.

To run topic modeling in Genim, we need to do a few things:
- Preprocess our text.
- Convert it into Gensim's bag-of-words format (this is different from the sparse matrices we get from scikit-learn).
- Run the topic model of our choice--we're going to use Latent Dirichlet Allocation (LDA).

Topic modeling tends to be _very_ data hungry, so we'll be using a much larger dataset than we've used for the previous notebooks.  Specifically: we'll grab the entirety of the Electronics reviews from Amazon.

import os
import pandas as pd

if not os.path.isfile("electronics_topic_modeling.parquet"):
    reviews = pd.read_json(
        "http://snap.stanford.edu/data/amazon/productGraph/categoryFiles/reviews_Electronics_5.json.gz",
        lines=True,
    )[["reviewText", "overall"]] # we'll use the "oveall" column in the next notebook
    reviews.to_parquet("electronics_topic_modeling.parquet")
else:
    reviews = pd.read_parquet("electronics_topic_modeling.parquet")
    
reviews = reviews["reviewText"].dropna().astype(str)
print(f"{reviews.shape[0]:,} reviews.")

We're going to employ some tricks to process this data as quickly as possible.  We're going to create our own preprocessing function using only some of the steps from Gensim's `preprocess_string` function.  Then we're going to call it in parallel over our texts using Joblib, for extra speed.

# Preprocessing
from joblib import Parallel, delayed
from gensim.parsing import preprocessing
from tqdm.notebook import tqdm

def preprocess(s):
    """Apply some of gensim's preprocessing tools."""
    s = preprocessing.strip_punctuation(s)
    s = preprocessing.strip_numeric(s)
    s = preprocessing.remove_stopwords(s.lower())
    s = preprocessing.strip_short(s)
    s = preprocessing.stem_text(s)
    return s.split()

parsed = Parallel(-1)(
    delayed(preprocess)(i)
    for i in tqdm(reviews, unit_scale=True, smoothing=0)
)

print(reviews.iloc[0])
print(parsed[0])

# Convert to a bag-of-words format using a Dictionary()
# object, which will handle filtering words for us and
# doing the bag of words conversion.
from gensim.corpora import Dictionary

# Dictonary([[word, word, ...], [word, word, ...], ...])
d = Dictionary(tqdm(parsed, desc="Creating gensim Dictionary", unit_scale=True))

# Remove rare and super common words
d.filter_extremes(no_above=0.5, no_below=20)

# Convert to bag of words
bow = [
    d.doc2bow(i)
    for i in tqdm(parsed, desc="Converting to bag of words", unit_scale=True)
]

# For topic modeling we usually only want texts of at least
# some minimum number of words, after filtering.  I'm picking
# 20, somewhat arbitrarily.
bow = [i for i in bow if len(i) >= 20]

print(f"{len(bow):,} reviews remain.")

Now, we build the LDA model on the bag of words representations.  Gensim has a few different topic models, but as mentioned, we're going to use one called Latent Dirichlet Allocation.  It tends to find more compact/coherent topics than some others, but its parameters can be very fiddly for datasets that aren't absurdly huge (this dataset is pretty moderately sized for LDA).

Gensim's LDA model does have an option to use "callbacks," which is a fancy way of saying "a function we pass that Gensim will periodically call for us."  Gensim typically calls these at the beginning and end of each pass over the dataset (LDA generally does several passes).  But, this makes it a bit hard to monitor the progress of each pass itself, so I'm going to slap together a quick class that'll print the progress on each iteration through a corpus.  Don't worry too much about the code details here.

class Corpus:
    def __init__(self, it):
        self.it = it
        self.n_passes = 1
        
    def __len__(self):
        return len(self.it)
    
    def __iter__(self):
        for i in tqdm(
            self.it,
            desc=f"Pass {self.n_passes}",
            unit_scale=True,
            smoothing=0,
        ):
            yield i
        self.n_passes += 1

# Set some environemnt variables before we run the model.
# Some of the lower-level libaries used by the LDA model
# are already multithreaded.  But we get more mileage
# out of multithreading at the LDA level.  So we need to
# single-thread the lower level libraries to avoid over-
# subscribing the CPU.
import os

os.environ["OMP_NUM_THREADS"] = "1"
os.environ["MLK_NUM_THREADS"] = "1"
os.environ["OPENBLAS_NUM_THREADS"] = "1"
os.environ["NUMEXPR_NUM_THREADS"] = "1"

from gensim.models.ldamulticore import LdaMulticore

lda = LdaMulticore(
    corpus=Corpus(bow),
    id2word=d,
    num_topics=10,
    passes=5, # 50 passes over the corpus
    workers=10,
)

print(lda.print_topics())

# Reshp that output into soemthing more readable.
for (topic_num, words) in lda.show_topics(formatted=False):
    print(f"TOPIC #{topic_num}")
    for w in words: 
        print(w)
    print()

We can obviously see-even though the mangling from the Porter Stemmer--some pretty consistent themes popping up in this data.  Cameras, keyboards and mice, headphones, speakers, etc. all show up pretty clearly in each topic.

We can also use the topic model to _transform_ a document into its topic distributions, i.e., into a vector that tells us "how much" of each topic is in the document.  I'll grab a review from Amazon (for the Logitech G610 Orion keyboard--a good, well priced mechanical keyboard if anyone happens to be curious).

sample_review = """Hello everyone this is my review on the Logitech G610 Orion Red. You can find the TL;DR at the bottom. :)

BUILD QUALITY:

The keyboard feels sturdy and has a bit of weight to it which results in a unit that feels robust and fairly high-quality. The top of the keyboard has a nice matte finish and the keycaps also have a matte, almost "satin" finish to them. The side of the keyboard is glossy and will surely attract fingerprints. Matte sides would have been much nicer and the decision to make them glossy seems a bit strange, but it does not seem to be that big a deal. The keycaps do leave a bit to be desired as they seem to be of a lower quality than the rest of the board and from the looks of it these keycaps will be glossing over fairly soon as I put this keyboard to extensive use. The media keys and volume roller situated in the upper right-hand corner of the keyboard both feel of a high quality.

KEYCAPS:

Do note, and this may be important for some folks, if you want to use custom keycaps, this keyboard does have an odd-size bottom row, particularly the spacebar which comes in at a very weird 5.75u. The entire bottom row on this keyboard is: 1.5u, 1.25u, 1.25u, 5.75u, 1.25u, 1.25u, 1.25u, 1.5u. With enough searching (and of course money) you could probably find keycaps for this bottom row, but do bear in mind that it may not be so easy and once you do find some nicer caps by time you pay for them you could be faced with a situation where the cost of those caps on top of the cost of this board exceeds that of getting a board with better keycaps to begin with. This is perhaps something for you to consider.

LIGHTING:

The lighting on the G610 only comes in white, however if you like the subdued styling of the keyboard but would prefer more lighting colors there is an RGB option, which is the Logitech G810. It is very similar to this board that in addition to the RGB lighting will feature Logitech's proprietary Romer-G switches as opposed to the Cherry brown or red switch options available on the G610. Whether one enjoys the feel of Romer-G switches is very much personal preference, but one thing to note with Romer-Gs is the backlighting will be more consistent across each key, unlike Cherry’s keys where thanks to the switch design influencing the placement of the LED, the lighting can be uneven across a key. For example, the bottom half of the keys of the number row (e.g. @, %, ^, &), and the secondary functions of the Numpad (e.g. Arrow Keys, END, PG DN, DEL) will be dimmer than the top half of the key. I have included a picture of a section the number pad demonstrating this. For many it may not be that big a deal, but if you are particular about things like that it may be an issue for you and could be one reason to look into Romer-G switches. Regardless of the inconsistent lighting, you should have no problem seeing everything in a dark room.

As far as lighting is concerned you can opt to use Logitech's software to customize between different effects such as "Breathing," "Star Effect," "Light Wave" and things like that or if you would rather not deal with the software you can forgo some of the fancier lighting schemes and simply hold down the keyboard's Brightness Setting key while pressing numbers 1, 2, 3, 4, 5, or 0 (0 is for solid) to cycle through different preset lighting schemes. It’s a pretty neat feature. The settings you select through Brightness key + numbers 1-5 and 0 should save through a cold boot. Considering my Logitech G600 mouse features a nice onboard memory that saves lighting and binding profiles, this keyboard could have certainly used something similar which for a unit that retails for $120 would have been nice to have.

FINAL THOUGHTS/OPINIONS:

This keyboard was a second choice for me as I intended to get a WASD CODE which retails for $150. The only reason I picked this Logitech up is the Cherry MX Red version was available on Amazon for $80 and I wanted to save a bit of money. Personally, I’d be reluctant to pay $100 for this keyboard and certainly not the retail asking price of $120 when for just $30 more I could have gotten a backlit board from WASD which has an even superior build quality, more choice of switches, and keycaps that are easier to replace... along with no "in your face" company logos, no media control switches (which for me are useless) and features onboard memory to easily save lighting profiles. Because of a lack of onboard memory to store lighting and key assignment profiles and because of the completely odd-ball and frankly quite annoying size of the bottom row’s spacebar which makes finding a full set of custom keycaps a nightmare, I will be removing one star.

In the end this is a nice keyboard for $80, and even while $100 would be doable it would be stretching it but I do trust the Logitech name and therefore can be pretty certain this keyboard will serve me well for the next few years, even if only as a backup. The flaws I mention are relatively minor and the quality and styling of this board far outweigh them. Personally I do not at all regret the purchase of this keyboard. If you want an easy, simple keyboard from a reputable company go ahead and pick this up and it should serve you well.

TL;DR

+Solid build quality
+Tried and true Cherry MX switches
+Backlit with five levels of brightness
+Easy to take apart (see youtube for teardowns)
+Comfortable to use

-Weird sizing for keycaps on bottom row, good luck finding a 5.75u spacebar :(
-No onboard memory to save lighting profile, key remaps, etc.
-A bit pricey considering no onboard memory, no USB passthrough, etc.
-Uneven lighting on some keys. While due to the design of Cherry switches this is not necessarily a fault of Logitech, the company could have opted to orient the keycap’s primary and secondary legends horizontally instead of vertically (e.g. as see on the backlit WASD Code) which would have provided more even lighting across the key.

Despite some flaws like a very odd sizing on the keycaps of the bottom row and lack of onboard memory to save lighting and other such profiles, The Logitech G610 Orion Red is solid board and one that I can happily recommend."""

# Run all the preprocessing
sample_review = preprocess(sample_review)
sample_review = d.doc2bow(sample_review)
# print(sample_review)

# Gensim's models override the `__get__()` method
# to transform data, so we can use this weird-looking
# syntax.  This filters out topics that are below a threshold.
# You can specify the threshold when creating the 
# LdaMulticore model--I left it at the default.
print(lda[sample_review])

# Or: get all topics, regardless of their proportion.
print(lda.get_document_topics(sample_review, minimum_probability=0))

# Convert to a Numpy array for, e.g., input to an ML
# or statistical model.
from gensim.matutils import corpus2dense
topics = lda.get_document_topics(sample_review, minimum_probability=0)
print(topics)

# corpus2dense expects a list of documents.
# `topics` is formatted like a single document.
# We also need to specify how many "terms" are in the 
# whole corpus.  corpus2dense returns an array with
# each document in a column, so we also transpose it.
print(corpus2dense([topics], 10).T)
